+++ tests/SimpleScannerTest.cpp
@@ -42,7 +42,21 @@ TEST (ScannerTest, scanner_ShouldReadCorrectlyTheListOfTokens) {
             // line 2
             OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::number_literal,
             OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::number_literal,
-            OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::endl_tk,
+            OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::number_literal,
+            OgreScriptLSP::endl_tk,
+            // line 3
+            OgreScriptLSP::endl_tk,
+            // line 4
+            OgreScriptLSP::endl_tk,
+            // line 5
+            OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::number_literal,
+            OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::number_literal,
+            OgreScriptLSP::number_literal, OgreScriptLSP::number_literal, OgreScriptLSP::number_literal,
+            OgreScriptLSP::endl_tk,
+    };
+    std::vector<std::string> numbers = {
+            ".34", "345.34", "234", "0.34", ".34", "-345.34", "-234", "-0.34", "-.34",
+            ".34f", "345.34f", "234f", "0.34f", ".34f", "-345.34f", "-234f", "-0.34f", "-.34f"
     };
 
     std::vector<OgreScriptLSP::TokenValue> result = scanner->parse();
@@ -53,6 +67,13 @@ TEST (ScannerTest, scanner_ShouldReadCorrectlyTheListOfTokens) {
     ASSERT_EQ(tokens.size(), result.size());
     for (int i = 0; i < tokens.size(); ++i) {
         ASSERT_EQ(tokens[i], result[i].tk);
+    }
+
+    // validate number literals
+    for (int i = 0, j = 0; i < tokens.size(); ++i) {
+        if (result[i].tk == OgreScriptLSP::number_literal) {
+            ASSERT_EQ(numbers[j++], result[i].literal);
+        }
     }
 }