+++ CHANGELOG.md
@@ -14,3 +14,4 @@ Fixed general issues related to the scripting language and not supported tokens.
 - Added: Support for missing <compute_program>|<compute_program_ref> blocks.
 - Added: Support for missing <shadow_receiver_material>|<shadow_caster_material> blocks.
 - Fixed: Error with characters allowed in identifier tokens.
+- Fixed: Issue with float numbers that ends with the character 'f'.